Refactor toss management for improved clarity and functionality
